Tests: Remove duplicate and near-duplicate test coverage
libeigen/eigen!3064

Closes #3167

+++ test/qtvector.cpp
@@ -137,5 +137,5 @@ EIGEN_DECLARE_TEST(qtvector) {
 
   // some Quaternion
   CALL_SUBTEST(check_qtvector_quaternion(Quaternionf()));
-  CALL_SUBTEST(check_qtvector_quaternion(Quaternionf()));
+  CALL_SUBTEST(check_qtvector_quaternion(Quaterniond()));
 }
